+++ gulpfile.js
@@ -34,7 +34,6 @@ const config = {
     'images': path.join('.', 'assets', 'images'),
     'bootstrap-sass': path.join('.', 'node_modules', 'bootstrap-sass'),
     'gill-sans': path.join('.', 'assets', 'webfonts', 'gill-sans'),
-    'academicons': path.join('.', 'assets', 'webfonts', 'academicons'),
     'open-sans': path.join('.', 'assets', 'webfonts', 'open-sans'),
     'bas-style-kit-js': path.join('.', 'bas-style-kit', '**/*.js'),
     'bootstrap-overrides-js': path.join('.', 'bootstrap-overrides', '**/*.js'),
@@ -103,7 +102,6 @@ gulp.task('minify--js', minifyJs);
 
 gulp.task('copy--font--gill-sans', copyFontGillSans);
 gulp.task('copy--font--open-sans', copyFontOpenSans);
-gulp.task('copy--font--academicons', copyFontAcademicons);
 
 gulp.task('copy--img--bas-logo', copyImagesBasLogo);
 gulp.task('copy--img--bas-roundel', copyImagesBasRoundel);
@@ -138,8 +136,7 @@ gulp.task('build--js', gulp.series(
 ));
 gulp.task('copy--fonts', gulp.parallel(
   'copy--font--gill-sans',
-  'copy--font--open-sans',
-  'copy--font--academicons'
+  'copy--font--open-sans'
 ));
 gulp.task('copy--img', gulp.parallel(
   'copy--img--bas-logo',
@@ -359,21 +356,6 @@ function copyFontOpenSans(done) {
     done
   );
 }
-
-function copyFontAcademicons(done) {
-  pump(
-    [
-      gulp.src([
-        path.join(config['sources']['academicons'], 'fonts', '*.eot'),
-        path.join(config['sources']['academicons'], 'fonts', '*.ttf'),
-        path.join(config['sources']['academicons'], 'fonts', '*.woff')
-      ]),
-      gulp.dest(path.join(config.destinations.dist, config.destinations.fonts, 'academicons'))
-    ],
-    done
-  );
-}
-
 
 function copyImagesBasLogo(done) {
   pump(